+++ Repositories/Interfaces/IRepository.cs
@@ -9,7 +9,7 @@ namespace AFBA.EPP.Repositories.Interfaces
     public interface IRepository<TEntity> where TEntity : class
     {
         Task<List<TEntity>> GetAll();
-        Task<TEntity> Get(int id);
+        Task<TEntity> Get(long id);
         Task<List<TEntity>> Find(Expression<Func<TEntity, bool>> predicate);
         Task<TEntity> SingleOrDefault(Expression<Func<TEntity, bool>> predicate);
         void Add(TEntity entity);
